chore: babel added for IE11 support
