+++ package.json
@@ -9,9 +9,12 @@
   "author": "Lindsey Wild",
   "license": "ISC",
   "devDependencies": {
+    "@babel/core": "^7.0.0-beta.55",
+    "babel-preset-env": "^1.7.0",
     "browser-sync": "^2.17.1",
     "del": "^3.0.0",
     "gulp": "^3.9.1",
+    "gulp-babel": "^8.0.0-beta.2",
     "gulp-clean-css": "^3.9.3",
     "gulp-inject-partials": "^1.0.4",
     "gulp-sass": "^2.3.2",
